Изменены типы данных

+++ src/main/java/com/spring/rest/system_bankpayments/repository/DepositAccountRepository.java
@@ -3,5 +3,5 @@ package com.spring.rest.system_bankpayments.repository;
 import com.spring.rest.system_bankpayments.entity.account.DepositAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepositAccountRepository extends JpaRepository<DepositAccount, Integer> {
+public interface DepositAccountRepository extends JpaRepository<DepositAccount, Long> {
 }
